fix: rename server-side upsert
